Fixed config file not getting generated on start Fixed loading dialog always displaying above all other applications Fixed issue where patching would fail if a game update introduced a new character/skin Added error dialog when the default nikke game data directory cannot be found

+++ NikkeModManagerCore/ModCollector.cs
@@ -7,6 +7,7 @@ using System.IO.Compression;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using NikkeModManagerCore.Exceptions;
 
 namespace NikkeModManagerCore {
     public class ModCollector {
@@ -94,6 +95,10 @@ namespace NikkeModManagerCore {
         private void BuildDefaultMod(string cacheDir, string defaultModDirectory, string gameDir) {
             NikkeMod mod = new NikkeDirectoryMod(gameDir);
             mod.Load(cacheDir);
+
+            if (mod.Bundles.Count == 0) {
+                throw new GameDataNotFoundException("Game Data Directory contains no asset bundles.");
+            }
 
             if (Directory.Exists(defaultModDirectory)) Directory.Delete(defaultModDirectory, true);
             Directory.CreateDirectory(defaultModDirectory);
